+++ lib/vgs_cg.h
@@ -55,7 +55,7 @@ typedef struct {
     uint32_t attr;        // Attribute
     uint32_t size;        // Size (0: 8x8, 1: 16x16, 2: 24x24, 3: 32x32 ... 31: 256x256)
     int32_t rotate;       // Rotate (-360 ~ 360)
-    uint32_t scale;       // Scale (0: disabled, or 1 ~ 400 percent)
+    uint32_t scale;       // Scale (0 ~ 3200 percent)
     uint32_t alpha;       // Alpha (0: disabled, or 0x000001 ~ 0xFFFFFF)
     uint32_t mask;        // Mask (0: disabled, or RGB888)
     uint32_t sly;         // Scale Lock (Y)